Bill of Materials
Revised Bill of Materials submission with information on parts from LCSC and digikey depending on availability.
I left resistors and capacitors blank except for the component value because we are all supposed to be using the 0805 components. I did not fill out LEDs because I do not need any particular LEDs. Connectors I was not sure how to source and how they compare to EAGLE components.