Updating UI solution
Rotary encoder interfacing is successful. Have yet to separate functionality into header and implementation files for lcd and rotary encoder functionality. The main program of this project is a test, however the main of my working project will have a similar outline to this; the difference mainly being that in the main switch case instead of incrementing a counter, registers in the AD9833 and MAX9939 will be set to adjust the signal, frequency, phase and gain as needed. HAVE NOT TESTED THE LCD PORTION OF CODE YET. NEED LCD/I2C backpack. It has worked in the past... probably effed something in it. It does build.