Functional user Interface program
UI code with 4x20 LCD and I2C backpack. Works well despite sometimes skipping modes when rotating encoder and sometimes missing button pushes. I tried incorporating a timer to check the push button at a set interval as well as an external interrupt, however this caused the LCD to turn on and off at regular intervals. I could not figure out why, so I scrapped that idea and just went with the regular polling as I had been doing. Modularized my program slightly by moving macros, globals and functions into header and implementation files for the rotary encoder and the user interface as a whole. The next step is to interface with the AD9833 signal generator chip.