Modularized AD9833 Header and Implementation
Refined the .h and .c file for AD9833 for my purposes. CURRENTLY UNTESTED. Need to test this solution with oscilloscope / logic analyzer when I have lab time on Wednesday. The solution builds, but who knows what it actually does. Had stuff working in lab last week so hopefully we didn't bust it up too much. Ideally, this can be integrated with the user interface. Next step is to implement signal amplification to the picture. If I use a strictly analog amplifier, presenting a voltage reading on the display might be more of a challenge. With the Programmable gain amplifier (PGA) it is probably a more straightforward equation based on register settings, signal voltage level estimate, and Vin of PGA.