Functional prototype code
This program is tested with the the Digilent oscilloscope feature. It effectively changes the frequency, phase and signal shape of the AD9833 via the user interface (rotary encoder).  I am realizing the rotary encoder is wildly, WILDLY inefficient for changing the frequency. A keypad would be a very handy option.  Additionally, I do not think i'll be able to display voltage readings on the LCD display.